# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden:
    <ul>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird entweder ein JSON-Array von JSON-Objekten, falls die Ressource eine Sammlung von Entitäten darstellt, oder ein JSON-Objekt, falls die Ressource eine einzelne Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Bei einer gültigen <code>DELETE</code>-Anfrage wird der Code <code>204</code> zurückgegeben.</b> Ist die Ressource nicht vorhanden, wird dies mit Code <code>404</code> belegt.</li>
...
      <li><b>Passwörter werden der Einfachheit halber im Klartext übergeben.</b></li>
...
    </ul>
...
servers:
...
paths:
  /veranstalter:
    post:
      tags:
        - Anwender
      summary: Fügt einen Veranstalter hinzu.
      description: >-
        Ein Veranstalter wird mit den entsprechenden Werten hinzugefügt. Ein zugeordneter User wird ebenfalls angelegt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/user'
                - $ref: '#/components/schemas/veranstalter'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
...
  /tickets:
    get:
      security:
        - basic authentication: []
      tags:
        - Authentifizierter und als Besucher authorisierter Anwender
      summary: Gibt die (eigenen) Tickets zurück.
      description: >-
        Alle (eigenen) Tickets werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: f_bezeichnung
          in: query
          description: 'Nur Tickets, deren zugeordnete Festivals die **Bezeichnung** `bezeichnung` enthalten, werden zurückgegeben.'
          schema:
            type: string
        - name: vip
          in: query
          description: 'Nur Tickets, die als **VIP-Kennzeichnung** `vip` besitzen, werden zurückgegeben.'
          schema:
            type: boolean
        - name: preis
          in: query
          description: 'Nur Tickets, deren **Preis** größer-gleich `preis` ist, werden zurückgegeben.'
          schema:
            type: number
            format: double
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        ticketid:
                          type: integer
                          description: '**RowID des Tickets**.'
                      required:
                        - ticketid
                    - $ref: '#/components/schemas/ticket'
  /orte:
    post:
      tags:
        - Authentifizierter und als Veranstalter authorisierter Anwender
      security:
        - basic authentication: []
      summary: Fügt einen Ort hinzu.
      description: >-
        Ein Ort wird mit den entsprechenden Werten hinzugefügt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/ort'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
...
  /tickets/{ticketid}:
    delete:
      security:
        - basic authentication: []
      tags:
        - Authentifizierter und als Besucher authorisierter Anwender
      summary: Löscht ein (eigenes) Ticket.
      description: >-
        Das (eigene) Ticket mit der **RowID** `ticketid` wird gelöscht.
      parameters:
        - name: ticketid
          in: path
          description: '**RowID des Tickets.**'
          schema:
            type: integer
          required: true
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
...
components:
  securitySchemes:
    basic authentication:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource.
          schema:
            type: string
    '204':
      description: '**No Content**.'
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
    '404':
      description: '**Not Found**.'
  schemas:
    user:
      type: object
      properties:
        email:
          type: string
          description: '**E-Mail-Adresse** des Users.'
        passwort:
          type: string
          description: '**Passwort** des Users.'
        vorname:
          type: string
          description: '**Vorname** des Users.'
        nachname:
          type: string
          description: '**Nachname** des Users.'
      required:
        - email
        - vorname
        - nachname
        - passwort
...
    veranstalter:
      type: object
      properties:
        veranstaltername:
          type: string
          description: '**Veranstaltername** des Veranstalters.'
      required:
        - veranstaltername
...
    ticket:
      type: object
      properties:
        preis:
          type: number
          format: double
          description: '**Preis** des Tickets (inkl. 2 Nachkommastellen).'
        datum:
          type: string
          format: date-time
          description: '**Datum** des Tickets im Format "YYYY-MM-DD" oder "YYYY-MM-DD HH:MM:SS".'
        vip:
          type: boolean
          description: '**VIP-Kennzeichnung** des Tickets.'
      required:
        - preis
        - datum
        - vip
...
    ort:
      type: object
      properties:
        bezeichnung:
          type: string
          description: '**Name** des Ortes.'
        land:
          type: string
          description: '**Land** des Orts.'
      required:
        - bezeichnung
        - land